Updated my files in all folders
I have not been working very hard on this in the past few months, but I am getting back into it.
This application started out as a way to catalog files in hanging file folders that are in various file drawers around the house. This allow, for example, looking for the location of a file by searching on a word, or part of a word.
Now, the project is expanding to include objects stored in drawers, on shelves, any place you have things you want to find later.