Store the images used in my web application. images
I am not a guru, nor do I have access to one. 

I am using the Apache httpd web server, and I am making up my own directory structure, so this will be in the file system as "/var/www/html/images", and on the local server as "http://localhost/dkfiles/images," even though users should not be allowed to navigate there. 

I have a lot to learn. Feel free to drop me a line with comments or suggestions.